Derive from QAbstractButton, fix focus after clicks
This fixes an issue where the widget does not have proper focus after
the widget was hidden in the mousePressEvent and later reshown, which
also results in hover state not being shown correctly.

Note that this issue seems to be caused solely by the use of
mousePressEvent, whereas mouseReleaseEvent (as is used by
QAbstractButton) seems to work fine.